move pbench_eval CLI scripts and helpers to examples/supercon-extraction
Relocates match.py, metrics.py, aggregate_accuracy_tokens.py, the
score_*_cli.py and generate_matches_cli.py scripts, plus
generate_gt_embeddings.py and generate_pred_embeddings.py into
examples/supercon-extraction as flat scripts, updates cross-imports to
local form, and drops the corresponding pyproject.toml entry points.
