+++ examples/supercon-extraction/generate_gt_embeddings.py
@@ -1,62 +1,72 @@
-"""Script to generate embeddings from ground truth property names.
+"""Generate embeddings for ground-truth property names from a HuggingFace dataset.
 
-Reference: https://ai.google.dev/gemini-api/docs/embeddings#task-types
+This script loads a ground-truth dataset from HuggingFace, extracts unique property
+names, generates embeddings using the Gemini embedding model, and saves them to JSON.
+
+Usage:
+    uv run python generate_gt_embeddings.py --hf_repo kilian-group/biosurfactants-extraction \
+        --hf_revision v0.0.0 --hf_split full
 """
 
-# standard imports
-import logging
-import pandas as pd
-from pathlib import Path
-from argparse import ArgumentParser
-from datasets import load_dataset
+import argparse
 import json
+import logging
+from pathlib import Path
 
-# third-party imports
+import pandas as pd
+from datasets import load_dataset
 from dotenv import load_dotenv
 from slugify import slugify
 
-# pbench imports
 import pbench
-from pbench_eval.match import generate_embeddings
 
-# local imports
-from utils import (
-    HF_DATASET_NAME,
-    HF_DATASET_REVISION,
-    HF_DATASET_SPLIT,
-)
+from match import generate_embeddings
 
-# Load env variables
 load_dotenv()
 
-parser = ArgumentParser(description="Generate embeddings from property names.")
+logger = logging.getLogger(__name__)
+
+parser = argparse.ArgumentParser(
+    description="Generate embeddings for ground-truth property names from a HuggingFace dataset."
+)
 parser = pbench.add_base_args(parser)
 args = parser.parse_args()
 pbench.setup_logging(args.log_level)
-logger = logging.getLogger(__name__)
 
-repo_name = args.hf_repo or HF_DATASET_NAME
-revision = args.hf_revision or HF_DATASET_REVISION
-split = args.hf_split or HF_DATASET_SPLIT
+if not args.hf_repo:
+    parser.error("--hf_repo is required")
+if not args.hf_revision:
+    parser.error("--hf_revision is required")
+if not args.hf_split:
+    parser.error("--hf_split is required")
+
+repo_name = args.hf_repo
+revision = args.hf_revision
+split = args.hf_split
 
 # Generate output path from repo_name, split, and revision
 output_filename = f"embeddings_{slugify(f'{repo_name}_{split}_{revision}')}.json"
 output_path = Path("scoring") / output_filename
 
-#
-# Load ground truth dataset
-#
+if output_path.exists() and not args.force:
+    logger.info(f"Output file already exists: {output_path}. Use --force to overwrite.")
+    raise SystemExit(0)
+
+logger.info(f"Loading dataset {repo_name} (split={split}, revision={revision})...")
 ds = load_dataset(repo_name, split=split, revision=revision)
 df = ds.to_pandas()
-# get unique property_names
+
+# Get unique property_names
 df = df.explode(column="properties").reset_index(drop=True)
 df = pd.concat(
     [df.drop(columns=["properties"]), pd.json_normalize(df["properties"])], axis=1
 )
 unique_property_names = df["property_name"].dropna().unique().tolist()
+logger.info(f"Found {len(unique_property_names)} unique property names")
 
+logger.info("Generating embeddings...")
 embeddings = generate_embeddings(unique_property_names)
-# save embeddings to parquet
+
 data = []
 for name, embedding in zip(unique_property_names, embeddings):
     data.append(
