+++ examples/supercon-extraction/generate_property_name_matches.py
@@ -32,7 +32,7 @@ from llm_utils import (
 # pbench imports
 from llm_utils.common import LLMChat
 import pbench
-from pbench_eval.match import generate_property_name_matches
+from match import generate_property_name_matches
 
 # local imports
 from utils import (
