+++ examples/supercon-extraction/score_recall_cli.py
@@ -32,13 +32,14 @@ import pandas as pd
 import logging
 
 import pbench
-from pbench_eval.metrics import compute_recall_per_material_property
 from pbench_eval.token_utils import (
     count_trials_per_group,
     count_zeroshot_trials_per_group,
 )
 from pbench_eval.stats import mean_sem_with_n
 from pbench_eval.cli_utils import load_rubric
+
+from metrics import compute_recall_per_material_property
 
 logger = logging.getLogger(__name__)
