+++ pyproject.toml
@@ -53,14 +53,6 @@ validator = [
 pbench-extract = "pbench.mass_extract_properties_from_llm:main"
 pbench-filter = "pbench.filter_candidates:main"
 pbench-eval = "pbench_eval.zeroshot:main"
-pbench-gt-embeddings = "pbench.generate_gt_embeddings:main"
-pbench-pred-embeddings = "pbench_eval.generate_pred_embeddings:main"
-pbench-generate-matches = "pbench_eval.generate_matches_cli:cli_main"
-pbench-score-precision = "pbench_eval.score_precision_cli:cli_main"
-pbench-score-recall = "pbench_eval.score_recall_cli:cli_main"
-pbench-score-f1 = "pbench_eval.score_f1_cli:cli_main"
-pbench-score-evidence = "pbench_eval.score_evidence_cli:cli_main"
-pbench-aggregate = "pbench_eval.aggregate_accuracy_tokens:cli_main"
 
 [tool.setuptools]
 package-dir = {"" = "src"}
