--- src/pbench_eval/generate_pred_embeddings.py
@@ -1,117 +1,0 @@
-"""Generate embeddings for predicted property names from Harbor jobs or CSV files.
-
-This script loads predictions from Harbor job directories or CSV files, generates
-embeddings for unique property names using the Gemini embedding model, and saves
-them to parquet files.
-
-Usage:
-    # From Harbor jobs directory
-    uv run pbench-pred-embeddings -jd JOBS_DIR -od OUTPUT_DIR
-
-    # From CSV files
-    uv run pbench-pred-embeddings -od OUTPUT_DIR -pd preds
-"""
-
-import argparse
-import logging
-
-import pandas as pd
-from dotenv import load_dotenv
-from slugify import slugify
-
-import pbench
-
-from pbench_eval.harbor_utils import get_harbor_data
-from pbench_eval.match import generate_embeddings
-
-logger = logging.getLogger(__name__)
-
-
-def main() -> None:
-    """CLI entry point."""
-    load_dotenv()
-
-    parser = argparse.ArgumentParser(
-        description="Generate embeddings for predicted property names."
-    )
-    parser = pbench.add_base_args(parser)
-    parser.add_argument(
-        "--agent",
-        "-a",
-        type=str,
-        default="gemini-cli",
-        help="Agent name used in Harbor jobs (default: gemini-cli)",
-    )
-    args = parser.parse_args()
-    pbench.setup_logging(args.log_level)
-
-    jobs_dir = args.jobs_dir
-    force = args.force
-    preds_dirname = args.preds_dirname
-
-    if jobs_dir is not None:
-        # Load predictions from Harbor jobs directory
-        df = get_harbor_data(jobs_dir)
-    else:
-        # Load predictions from CSV files
-        if args.output_dir is None:
-            parser.error("--output_dir is required when not using --jobs_dir")
-        preds_dir = args.output_dir / preds_dirname
-        preds_files = list(preds_dir.glob("*.csv"))
-        if not preds_files:
-            raise FileNotFoundError(f"No CSV files found in {preds_dir}")
-        dfs = []
-        for file in preds_files:
-            file_df = pd.read_csv(file)
-            assert "refno" in file_df.columns, (
-                "refno column not found in predictions CSV"
-            )
-            assert file_df["refno"].nunique() == 1, (
-                "Expected only one unique refno per predictions CSV"
-            )
-            dfs.append(file_df)
-        df = pd.concat(dfs, ignore_index=True)
-
-    if args.output_dir is None:
-        parser.error("--output_dir is required")
-
-    embeddings_dir = args.output_dir / "pred_embeddings"
-    embeddings_dir.mkdir(parents=True, exist_ok=True)
-
-    for (agent, model, refno), group in df.groupby(
-        ["agent", "model", "refno"], dropna=False
-    ):
-        save_path = (
-            embeddings_dir / f"{slugify(agent)}_{slugify(model)}_{refno}.parquet"
-        )
-        if save_path.exists() and not force:
-            logger.info(
-                f"Embeddings already exist for {agent=} {model=} {refno=}, skipping..."
-            )
-            continue
-
-        logger.info(f"Generating embeddings for {agent=} {model=} {refno=}...")
-        preds_df = group[group["refno"] == refno]
-        property_names = preds_df["property_name"].dropna().tolist()
-        # get unique property names
-        unique_property_names = list(set(property_names))
-        # generate embeddings
-        embeddings = generate_embeddings(unique_property_names)
-        # save embeddings to parquet
-        assert len(preds_df["refno"].unique()) == 1, "Expected only one refno per file"
-        refno = preds_df["refno"].unique()[0]
-        embeddings_df = pd.DataFrame(
-            {
-                "refno": [refno] * len(unique_property_names),
-                "property_name": unique_property_names,
-                "embedding": embeddings,
-                "agent": [agent] * len(unique_property_names),
-                "model": [model] * len(unique_property_names),
-            }
-        )
-        embeddings_df.to_parquet(save_path)
-        logger.info(f"Saved embeddings to {save_path} with {len(embeddings_df)} rows")
-
-
-if __name__ == "__main__":
-    main()
